reorganize client side schema/type definitions
