+++ client/src/components/ManifestForm/ContactForm/ContactForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
 import { Col, Row } from 'react-bootstrap';
-import { Manifest } from 'types';
+import { Manifest } from 'types/Manifest';
 import { HtForm } from 'components/Ht';
 
 interface ContactFormProps {
