+++ client/src/components/ManifestForm/HandlerSearch/TransporterSearchForm.tsx
@@ -4,9 +4,8 @@ import React, { useEffect, useState } from 'react';
 import { Button, Col, Row } from 'react-bootstrap';
 import { SubmitHandler, UseFieldArrayAppend, useForm } from 'react-hook-form';
 import htApi from 'services';
-import { Handler, Manifest } from 'types';
-import { HandlerType } from 'types/Handler/Handler';
-import { Transporter } from 'types/Handler/Transporter';
+import { Manifest } from 'types/Manifest';
+import { Handler, HandlerType, Transporter } from 'types/Handler';
 
 interface Props {
   handleClose: () => void;
