+++ client/src/components/ManifestForm/ManifestForm.tsx
@@ -11,9 +11,8 @@ import { Button, Col, Row } from 'react-bootstrap';
 import { FormProvider, SubmitHandler, useFieldArray, useForm } from 'react-hook-form';
 import htApi from 'services';
 import { addMsg, useAppDispatch } from 'store';
-import { Handler, Manifest } from 'types';
-import { Transporter } from 'types/Handler';
-import { HandlerType, ManifestHandler } from 'types/Handler/Handler';
+import { Manifest } from 'types/Manifest';
+import { Transporter, HandlerType, ManifestHandler } from 'types/Handler';
 import { WasteLine } from 'types/WasteLine';
 import HandlerForm from './HandlerForm';
 import AddTsdf from './Tsdf';
