+++ client/src/components/ManifestForm/Transporter/AddTransporter.tsx
@@ -3,7 +3,8 @@ import TransporterSearchForm from 'components/ManifestForm/HandlerSearch/Transpo
 import React from 'react';
 import { Col, Row } from 'react-bootstrap';
 import { UseFieldArrayAppend } from 'react-hook-form';
-import { Handler, Manifest } from 'types';
+import { Handler } from 'types/Handler';
+import { Manifest } from 'types/Manifest';
 
 interface Props {
   handleClose: () => void;
@@ -12,6 +13,14 @@ interface Props {
   appendTransporter: UseFieldArrayAppend<Manifest, 'transporters'>;
 }
 
+/**
+ * Returns a modal for adding a transporter to a manifest
+ * @param handleClose
+ * @param show
+ * @param appendTransporter
+ * @param currentTransporters
+ * @constructor
+ */
 function AddTransporter({ handleClose, show, appendTransporter, currentTransporters }: Props) {
   return (
     <HtModal showModal={show ? show : false} handleClose={handleClose}>
