+++ client/src/components/ManifestForm/WasteLine/WasteLineForm.tsx
@@ -4,7 +4,7 @@ import HazardousWasteForm from 'components/ManifestForm/WasteLine/HazardousWaste
 import React from 'react';
 import { Button, Container, Row } from 'react-bootstrap';
 import { FormProvider, UseFieldArrayAppend, useForm } from 'react-hook-form';
-import { Manifest } from 'types';
+import { Manifest } from 'types/Manifest';
 import { WasteLine } from 'types/WasteLine';
 import QuantityForm from './QuantityForm';
