+++ client/src/features/manifest/ManifestDetails.tsx
@@ -4,7 +4,7 @@ import useHtAPI from 'hooks/useHtAPI';
 import useTitle from 'hooks/useTitle';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { Manifest } from 'types';
+import { Manifest } from 'types/Manifest';
 
 function ManifestDetails() {
   const { mtn, action } = useParams();
