+++ client/src/types/Handler/Contact.ts
@@ -1,0 +1,13 @@
+export interface Contact {
+  firstName?: string;
+  middleInitial?: string;
+  lastName?: string;
+  phone?: Phone;
+  email?: string;
+  companyName?: string;
+}
+
+export interface Phone {
+  number: string;
+  extension?: string;
+}
