--- client/src/types/Handler/Transporter.ts
@@ -1,6 +1,0 @@
-import { Handler } from 'types/index';
-
-export interface Transporter extends Handler {
-  order: number;
-  manifest?: number;
-}
