+++ client/src/types/Handler/address.ts
@@ -1,50 +1,3 @@
-import { Contact, Phone } from 'types/Manifest/Contact';
-
-export interface Handler {
-  name: string;
-  epaSiteId: string;
-  mailingAddress: Address;
-  siteAddress: Address;
-  contact: Contact;
-  emergencyPhone: Phone;
-  order?: number;
-  registered?: boolean;
-  modified?: boolean;
-  limitedEsign?: boolean;
-  canEsign?: boolean;
-  siteState?: Locality;
-  hasRegisteredEmanifestUser?: boolean;
-  gisPrimary?: boolean;
-}
-
-export interface ManifestHandler extends Handler {
-  // paperSignatureInfo: PaperSignature
-  electronicSignatureInfo?: Array<ElectronicSignature>;
-}
-
-export interface ElectronicSignature {
-  signer?: Signer;
-  signatureDate?: string;
-}
-
-export interface Signer {
-  firstName?: string;
-  middleInitial?: string;
-  lastName?: string;
-  companyName?: string;
-}
-
-/**
- * Used to specify whether a handler is a generator, transporter, or tsdf.
- */
-export enum HandlerType {
-  Generator = 'generator',
-  Tsd = 'designatedFacility',
-  Transporter = 'transporter',
-}
-
-export type HandlerTypeString = `${HandlerType}`;
-
 /**
  * Used to specify whether a siteAddress or mailingAddress
  */
