--- client/src/types/WasteLine/index.ts
@@ -1,8 +1,0 @@
-import {
-  ContainerType,
-  QuantityCode,
-  QuantityDescription,
-  WasteLine,
-} from 'types/WasteLine/WasteLine';
-
-export type { WasteLine, ContainerType, QuantityCode, QuantityDescription };
