--- client/src/types/index.ts
@@ -1,5 +1,0 @@
-import { Manifest } from './Manifest/';
-import { Handler } from './Handler';
-import { usePaginationProps } from './hooks';
-
-export type { Manifest, Handler, usePaginationProps };
