+++ client/src/types/store.ts
@@ -1,5 +1,5 @@
 /**
- * The stored information on the current haztrak user
+ * The Redux stored information on the current haztrak user
  */
 export interface UserState {
   user?: string;
@@ -9,7 +9,8 @@ export interface UserState {
 }
 
 /**
- * Encompasses all global notifications
+ * Schema of a user's alerts stored in the Redux store
+ * Note, we don't store these in the back end database, this is just for looks.
  */
 export interface NotificationState {
   alert: Array<Alert>;
@@ -28,19 +29,22 @@ export interface Alert {
 }
 
 /**
- * The stored information on the current haztrak user
+ * The user's RCRAInfo account data stored in the Redux store
  */
 export interface RcraProfileState {
   user: string | undefined;
   rcraAPIID?: string;
-  rcraAPIKey?: string;
   rcraUsername?: string;
+  rcraAPIKey?: string;
   epaSites?: Record<string, ProfileEpaSite>;
   phoneNumber?: string;
   loading?: boolean;
   error?: string;
 }
 
+/**
+ * The user's site permissions for an EPA site in RCRAInfo
+ */
 export interface ProfileEpaSite {
   epaId: string;
   permissions: {
